+++ source/Gui.cpp
@@ -241,12 +241,10 @@ bool Gui::miniMapclicked(const CEGUI::EventArgs& e)
 {
     CEGUI::MouseEventArgs& ee = (CEGUI::MouseEventArgs&)e;
 
-    Ogre::Vector2 cc = ODFrameListener::getSingleton().getGameMap()->getMiniMap()->camera_2dPositionFromClick(ee.position.d_x,
-                                                                                                              ee.position.d_y );
+    Ogre::Vector2 cc = ODFrameListener::getSingleton().getGameMap()->getMiniMap()->camera_2dPositionFromClick((int)ee.position.d_x,
+                                                                                                              (int)ee.position.d_y);
 
     modeManager->mMc->frameListener->cm->onMiniMapClick(cc);
-    // ODFrameListener::getSingleton().getGameMap()->getMiniMap()->draw();
-    // ODFrameListener::getSingleton().getGameMap()->getMiniMap()->swap();
 
     //std::cerr<< xx <<" "<< yy << " " <<std::endl;
     return true;
